allow running tests but have consistent GETH node running locally

+++ scripts/test_wrapper.sh
@@ -4,15 +4,24 @@ set -a
 # Setup
 local_dir="$(dirname "$0")"
 PARENT_DIR=`pwd`
+GETH_DOCKER_CONTAINER_NAME="geth-node"
+CONTAINER_WAS_RUNNING=`docker ps -a --filter "name=${DOCKER_CONTAINER_NAME}" --filter "status=running" --quiet`
+
 
 ################# Run Dependencies #########################
 for path in ${local_dir}/test-dependencies/*; do
     [ -d "${path}" ] || continue # if not a directory, skip
     source "${path}/env_vars.sh" # Every dependency should have env_vars.sh + run.sh executable files
-    echo "Executing [${path}/run.sh]"
-    ${path}/run.sh
-    if [ $? -ne 0 ]; then
-      exit 1
+
+    if [ -z ${CONTAINER_WAS_RUNNING} ]; then
+        echo "Container ${GETH_DOCKER_CONTAINER_NAME} is not currently running"
+        echo "Executing [${path}/run.sh]"
+        ${path}/run.sh
+        if [ $? -ne 0 ]; then
+          exit 1
+        fi
+    else
+      echo "Container ${GETH_DOCKER_CONTAINER_NAME} is already running. Not starting again."
     fi
 done
 ############################################################
@@ -44,14 +53,16 @@ fi
 ############################################################
 
 ################# CleanUp ##################################
-echo "Bringing GETH Daemon Down"
-docker rm -f geth-node
+if [ ! -z ${CONTAINER_WAS_RUNNING} ]; then
+    echo "Bringing GETH Daemon Down"
+    docker rm -f geth-node
 
-# Cleaning extra DAG file, so we do not cache it - travis
-if [[ "X${RUN_CONTEXT}" == "Xtravis" ]];
-then
-  new_dag=`ls -ltr $DATA_DIR/.ethash/* | tail -1 | awk '{print $9}' | tr -d '\n'`
-  rm -Rf $new_dag
+    # Cleaning extra DAG file, so we do not cache it - travis
+    if [[ "X${RUN_CONTEXT}" == "Xtravis" ]];
+    then
+      new_dag=`ls -ltr $DATA_DIR/.ethash/* | tail -1 | awk '{print $9}' | tr -d '\n'`
+      rm -Rf $new_dag
+    fi
 fi
 ############################################################
